refactor: renaming NativeExplorerAPI to EtherscanAPI
