+++ packages/web3-providers/src/Etherscan/apis/ExplorerAPI.ts
@@ -1,11 +1,11 @@
 import urlcat from 'urlcat'
 import { ChainId, ExplorerURL } from '@masknet/web3-shared-evm'
-import type { Transaction } from './types.js'
-import { toTransaction } from './helpers.js'
-import type { ExplorerAPI } from '../entry-types.js'
-import { fetchJSON } from '../entry-helpers.js'
+import type { Transaction } from '../types.js'
+import { toTransaction } from '../helpers.js'
+import type { ExplorerAPI } from '../../entry-types.js'
+import { fetchJSON } from '../../entry-helpers.js'
 
-export class NativeExplorerAPI implements ExplorerAPI.Provider {
+export class EtherscanExplorerAPI implements ExplorerAPI.Provider {
     async getLatestTransactions(
         chainId: ChainId,
         account: string,
